+++ swift_book_pdf/cli.py
@@ -188,6 +188,12 @@ def cli() -> None:
     default=None,
     help='When generating an EPUB, override the version number. Include "beta" for beta versions. If not provided, the version will be determined by parsing the table of contents.',
 )
+@click.option(
+    "--publisher",
+    type=str,
+    default=None,
+    help="When generating an EPUB, set the publisher metadata field to the specified value. If not provided, the publisher will not be set.",
+)
 @click.option("--verbose", is_flag=True, help="Enable verbose logging.")
 @click.version_option(
     prog_name="Swift-Book-PDF",
@@ -212,6 +218,7 @@ def run(  # noqa: PLR0913
     export_cover_image: bool,
     cover_footer_line: str | None,
     override_version: str | None,
+    publisher: str | None,
 ) -> None:
     configure_logging(verbose)
     logger = logging.getLogger(__name__)
@@ -230,6 +237,7 @@ def run(  # noqa: PLR0913
         export_cover_image=export_cover_image,
         cover_footer_line=cover_footer_line,
         override_version=override_version,
+        publisher=publisher,
     )
 
     try:
@@ -294,6 +302,7 @@ def _build_config(
         export_cover_image=options.export_cover_image,
         cover_footer_line=options.cover_footer_line,
         override_version=options.override_version,
+        publisher=options.publisher,
     )